Make json serialising more informative (add offsets info)
Events now can be represented by its offsets in original sentence.
But it seems there is some bug in OpenIE - some offsets from Action
part of extraction are not in the output, so we loss part of
extracted Action.
The whole thing is for experimenting with some kinds of DL models.